# openapi.yml
openapi: 3.0.3
info:
  title: ASPB Autowebinar API
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/v1/creator/webinars/{webinarId}/uploads:
    post:
      summary: Create a private resumable multipart video upload
      security:
        - userSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/WebinarId'
        - name: Idempotency-Key
          in: header
          required: true
          description: Stable per logical upload request. A replay with the same request contract returns the existing upload and freshly signed operations only for missing parts; a conflicting contract is rejected.
          schema: { type: string, minLength: 8, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [fileName, mimeType, sizeBytes]
              properties:
                fileName: { type: string, maxLength: 240 }
                mimeType: { type: string, enum: [video/mp4, video/quicktime, video/webm] }
                sizeBytes: { type: string, pattern: '^\\d{1,13}$' }
                checksumSha256: { type: string, pattern: '^[0-9a-f]{64}$' }
      responses:
        '200':
          {
            description: Idempotent replay returned the existing asset/upload and fresh operations only for missing parts,
          }
        '201':
          {
            description: Asset and temporary provider or same-origin part operations created; origin key is not returned,
          }
        '400': { description: Invalid type, size or request contract }
        '404': { description: Webinar is unavailable in the active tenant }
        '409': { description: Idempotency-Key was already used with a different server-owned upload contract }
        '503': { description: Private object storage adapter is not configured }
  /api/v1/creator/uploads/{uploadId}/complete:
    post:
      summary: Idempotently complete a private multipart upload and enqueue validation
      description: Replays a provider-committed completion safely by verifying the object metadata before persisting one deduplicated job.
      security:
        - userSessionCookie: []
      parameters:
        - name: uploadId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [parts]
              properties:
                parts:
                  type: array
                  minItems: 1
                  maxItems: 1000
                  items:
                    type: object
                    additionalProperties: false
                    required: [partNumber, etag]
                    properties:
                      partNumber: { type: integer, minimum: 1, maximum: 10000 }
                      etag: { type: string, minLength: 1, maxLength: 256 }
      responses:
        '200': { description: Upload completed and one deduplicated media job exists }
        '404': { description: Upload is unavailable in the active tenant }
        '409':
          { description: Client, server checkpoint and provider ListParts state disagree; completion was not attempted }
        '422': { description: Trusted signature, MIME, integrity or duration validation failed }
  /api/v1/creator/uploads/{uploadId}/parts:
    post:
      summary: Idempotently record one directly uploaded multipart part
      description: Stores only bounded part metadata. Provider size/checksum are reconciled through ListParts before resume/complete. External storage bytes bypass Express; self-hosted bytes use the separately authenticated streaming endpoint.
      security:
        - userSessionCookie: []
      parameters:
        - name: uploadId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [partNumber, etag]
              properties:
                partNumber: { type: integer, minimum: 1, maximum: 10000 }
                etag: { type: string, minLength: 1, maxLength: 256 }
      responses:
        '200': { description: Part checkpoint recorded and safe progress returned }
        '404': { description: Upload is unavailable in the active tenant }
        '409':
          {
            description: Upload is expired/completed/cancelled,
            or the same part number conflicts with the provider/server checkpoint,
          }
  /api/v1/creator/uploads/{uploadId}/parts/{partNumber}/content:
    put:
      summary: Stream one self-hosted multipart part into private persistent storage
      description: Requires the platform session, active tenant membership, creator-to-Webinar relation and double-submit CSRF token. Content-Length and MIME must match server-owned upload metadata. The path never grants access by itself and never exposes a filesystem path or storage key.
      security:
        - userSessionCookie: []
      parameters:
        - name: uploadId
          in: path
          required: true
          schema: { type: string }
        - name: partNumber
          in: path
          required: true
          schema: { type: integer, minimum: 1, maximum: 10000 }
        - name: x-csrf-token
          in: header
          required: true
          schema: { type: string }
        - name: Content-Length
          in: header
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          video/mp4:
            schema: { type: string, format: binary }
          video/quicktime:
            schema: { type: string, format: binary }
          video/webm:
            schema: { type: string, format: binary }
      responses:
        '200': { description: Part fsynced, SHA-256 ETag returned and server checkpoint committed }
        '400': { description: Part number, MIME or exact byte length does not match server metadata }
        '401': { description: Platform session is absent or invalid }
        '403': { description: CSRF token is missing or invalid }
        '404': { description: Upload is unknown, cross-tenant or unrelated to this author }
        '409': { description: Upload is inactive or this part already has different content }
        '411': { description: Content-Length is required }
        '415': { description: MIME is missing or content encoding is unsupported }
        '429': { description: Self-hosted part rate limit exceeded }
        '503': { description: Private media storage is unavailable }
  /api/v1/creator/uploads/{uploadId}/resume:
    post:
      summary: Resume a multipart upload without retransmitting completed parts
      description: Returns fresh temporary operations only for missing parts and never returns the storage key. If the provider already committed completion, trusted server checkpoints are returned so the client can replay complete safely.
      security:
        - userSessionCookie: []
      parameters:
        - name: uploadId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
      responses:
        '200': { description: Completed checkpoints and newly signed missing-part operations returned }
        '404': { description: Upload is unavailable in the active tenant }
        '409':
          {
            description: Upload is expired/completed/cancelled,
            belongs to another provider,
            or provider/server checkpoints disagree,
          }
  /api/v1/creator/media/{assetId}/status:
    get:
      summary: Read safe tenant-scoped media processing status
      security:
        - userSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200': { description: Safe status without storage keys or origin URLs }
        '404': { description: Asset is unavailable in the active tenant }
...
  /api/v1/creator/webinars/{webinarId}/materials:
    get:
      summary: List private materials for one creator Webinar
      security: [{ userSessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/WebinarId' }]
      responses:
        '200': { description: Private no-store metadata without storage authority }
        '404': { description: Unknown, cross-tenant or unauthorized Webinar }
  /api/v1/creator/webinars/{webinarId}/materials/uploads:
    post:
      summary: Start an idempotent multipart upload for an allowlisted private material
      description: Binds tenant, Webinar, name, MIME, exact size and optional SHA-256. Maximum size is 100 MiB; storage authority is short-lived.
      security: [{ userSessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/WebinarId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MaterialUploadCreateRequest' } } }
      responses:
        '201': { description: Upload intent and short-lived part operations created }
        '200': { description: Exact idempotent upload intent replayed }
        '404': { description: Unknown, cross-tenant or unauthorized Webinar }
        '409': { description: Idempotency conflict or Webinar is not editable }
        '413': { description: Material exceeds 100 MiB }
        '415': { description: Extension/MIME combination is not allowlisted }
  /api/v1/creator/material-uploads/{uploadId}/parts:
    post:
      summary: Record one provider-completed multipart part
      security: [{ userSessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/UploadId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MultipartPartRequest' } } }
      responses:
        '200': { description: Part receipt recorded idempotently }
        '404': { description: Unknown, expired, foreign or unauthorized upload }
        '409': { description: Part receipt conflicts with the existing binding }
  /api/v1/creator/material-uploads/{uploadId}/complete:
    post:
      summary: Finalize after exact provider size, MIME and checksum verification
      security: [{ userSessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/UploadId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [parts]
              properties:
                parts:
                  {
                    type: array,
                    minItems: 1,
                    maxItems: 1000,
                    items: { $ref: '#/components/schemas/MultipartPartRequest' },
                  }
      responses:
        '200': { description: Material ready or exact completion replayed idempotently }
        '404': { description: Unknown, expired, foreign or unauthorized upload }
        '409': { description: Parts, binding or integrity verification failed }
  /api/v1/creator/material-uploads/{uploadId}/resume:
    post:
      summary: Reconcile parts and issue short-lived operations for missing parts
      security: [{ userSessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/UploadId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        content: { application/json: { schema: { type: object, additionalProperties: false } } }
      responses:
        '200': { description: Private no-store resume projection }
        '404': { description: Unknown, expired, foreign or unauthorized upload }
  /api/v1/creator/materials/{materialId}:
    delete:
      summary: Soft-delete one material using optimistic revision
      description: Physical provider deletion remains governed by the separately approved retention policy.
      security: [{ userSessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/MaterialId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RevisionRequest' } } }
      responses:
        '200': { description: Material unavailable and audited }
        '404': { description: Unknown, cross-tenant or unauthorized material }
        '409': { description: Material revision is stale }
...
components:
  parameters:
    AssetId:
      name: assetId
      in: path
      required: true
      schema:
        type: string
...
    WebinarId:
      name: webinarId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 191
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 191
        pattern: '^[A-Za-z0-9._:-]+$'
    CsrfToken:
      name: x-csrf-token
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 4096 }
...
    MaterialId:
      name: materialId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 191 }
    UploadId:
      name: uploadId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 191 }
...
  schemas:
    RevisionRequest:
      type: object
      additionalProperties: false
      required: [expectedRevision]
      properties: { expectedRevision: { type: integer, minimum: 1 } }
...
    MaterialUploadCreateRequest:
      type: object
      additionalProperties: false
      required: [displayName, fileName, mimeType, sizeBytes]
      properties:
        displayName: { type: string, minLength: 2, maxLength: 240 }
        fileName: { type: string, minLength: 5, maxLength: 240 }
        mimeType: { type: string, minLength: 3, maxLength: 160 }
        sizeBytes:
          oneOf:
            - { type: integer, minimum: 1, maximum: 104857600 }
            - { type: string, pattern: '^\d{1,12}$' }
        checksumSha256: { type: string, pattern: '^[0-9a-f]{64}$' }
    MultipartPartRequest:
      type: object
      additionalProperties: false
      required: [partNumber, etag]
      properties:
        partNumber: { type: integer, minimum: 1, maximum: 1000 }
        etag: { type: string, minLength: 1, maxLength: 256 }
...
  securitySchemes:
    metricsBearer:
      type: http
      scheme: bearer
    userSessionCookie:
      type: apiKey
      in: cookie
      name: aspb_user_session
    participantSessionCookie:
      type: apiKey
      in: cookie
      name: aspb_room_token
    adminSessionCookie:
      type: apiKey
      in: cookie
      name: aspb_admin_session